maybe have fixed TMM method, let's see if it passes
